Add local TTL option to dnsmasq
-T, --local-ttl=<time>
    When replying with information from /etc/hosts or the DHCP leases
file dnsmasq by default sets the time-to-live field to zero, meaning
that the requestor should not itself cache the information. This is
the correct thing to do in almost all situations. This option allows a
time-to-live (in seconds) to be given for these replies. This will
reduce the load on the server at the expense of clients using stale
data under some circumstances.

[jow: change -T to --local--ttl to conform with the other options]

Signed-off-by: Andrew Byrne <openwrt@andy.id.au>

openwrt-devel mailing list
openwrt-devel@lists.openwrt.org
https://lists.openwrt.org/mailman/listinfo/openwrt-devel