[package] Snort: update to 2.9.2.2
I've managed to update Snort to version 2.9.2.2. I did not want to split this patch even if it has so many deltas - there are many deletions of unused patches. Also there is a configuration file that has almost the same functionality as the 2.8.4.1 one. The reason I've not written a patch for update to 2.9.3 is to preserve compatibility with snortsam which has not patch for 2.9.3 yet.

If this patch becomes accepted then patch http://patchwork.openwrt.org/patch/2381/ becomes obsolete, because it depends on 2.8.4.1. Next patch (3/3) introduces support for Snortsam package (Snortsam is snort version independent, just needs snort to be patched with proper patch).

Signed-off-by: Jiri Slachta <jiri@slachta.eu>