[brcm47xx]: deactivate mips wait instruction for MIPS 74K on Broadcom SoCs
When not removing support for the wait instruction, when ruining on the 
BCM4706 SoC the system stops after msleep() without an exception. I 
haven't seen this problem on the other Broadcom SoCs with a 74K CPU, 
but in the Broadcom SDK it is deactivated for all CPUs of this type.